apps/twitter/twitter-bot/requirements.txt
tweepy==4.14.0
openai==1.51.0
python-dotenv==1.0.1
